refactor: move threads into own directory
