+++ src/node-lib/curriculum-api-2023/queries/unitListing/threads/getAllLearningThemes.ts
@@ -1,0 +1,17 @@
+import { UnitData } from "../unitListing.schema";
+
+import { LearningThemes } from "./threads.schema";
+
+export const getAllLearningThemes = (units: Array<Array<UnitData>>) => {
+  return units
+    .flat()
+    .reduce((acc, unit) => {
+      unit.learningThemes?.forEach((theme) => {
+        if (!acc.find((t) => t.themeSlug === theme.themeSlug)) {
+          acc.push(theme);
+        }
+      });
+      return acc;
+    }, [] as LearningThemes)
+    .sort((a, b) => (a.themeTitle > b.themeTitle ? 1 : -1));
+};
