+++ src/node-lib/curriculum-api-2023/queries/unitListing/threads/threads.schema.ts
@@ -1,0 +1,23 @@
+import { z } from "zod";
+
+export const threadsResponseSchema = z.array(
+  z.object({
+    threads: z.array(
+      z.object({
+        threads: z
+          .array(z.object({ theme_slug: z.string(), theme_title: z.string() }))
+          .nullish(),
+        unit_id: z.number(),
+      }),
+    ),
+  }),
+);
+
+export const learningThemesSchema = z.object({
+  themeTitle: z.string(),
+  themeSlug: z.string(),
+});
+
+const learningThemes = z.array(learningThemesSchema);
+
+export type LearningThemes = z.infer<typeof learningThemes>;
